backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
google-generativeai==0.3.2
chromadb==0.4.18
PyPDF2==3.0.1
python-docx==1.1.0
langchain==0.0.350
langchain-google-genai==0.0.5
pydantic==2.5.0
pydantic[email]==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiofiles==23.2.1
httpx==0.25.2
numpy==1.24.3
sentence-transformers==2.2.2 
chromadb
PyJWT==2.8.0
bcrypt==4.1.2
motor==3.3.2
pymongo==4.6.0
